chore: adjust pages table numbers type

+++ priv/repo/migrations/20240619001632_add_pages_table.exs
@@ -6,8 +6,8 @@ defmodule EtlChallenge.Repo.Migrations.AddPagesTable do
       add :page, :integer, primary_key: true
       add :is_failed, :boolean, default: false
       add :fail_reason, :string
-      add :numbers, {:array, :decimal}, default: []
-      add :sorted_numbers, {:array, :decimal}, default: []
+      add :numbers, {:array, :float}, default: []
+      add :sorted_numbers, {:array, :float}, default: []
       add :last_fetched_at, :utc_datetime
 
       timestamps()
